requirements.txt
Django==5.0
djangorestframework>=3.12
djangorestframework-simplejwt==5.3.1
psycopg2-binary>=2.9
celery>=5.6
django-celery-beat>=2.8.1
redis>=5.0
drf-yasg>=1.21
requests>=2.31
python-dotenv>=1.1
pytz>=2023.3
django-cors-headers>=3.14

# Development tools
black>=23.0
flake8>=6.0
isort>=5.12

